## Before
```
Completed 7,421 operations in a weighted-average of 5.00s (1,482.92 ops/s, 0.001 s/op)
```

| Test       | Arguments                                                | 12.11.0 | source  | %Change |
|------------|----------------------------------------------------------|---------|---------|---------|
| download   | --size 10240 --parallel 64                               | 1991.58 | 1911.60 | -4.02%  |
| download   | --size 10485760 --parallel 32                            | 144.92  | 145.36  | 0.30%   |
| download   | --size 10485760 --parallel 32 --sync                     | -1.00   | -1.00   | 0.00%   |
| download   | --size 1073741824 --parallel 1 --warmup 60 --duration 60 | 0.17    | 0.18    | 5.88%   |
| download   | --size 1073741824 --parallel 8 --warmup 60 --duration 60 | 1.24    | 1.25    | 0.81%   |


## After
```
Completed 7,392 operations in a weighted-average of 4.999s (1,479 ops/s, 0.0006763 s/op)
```

| Test       | Arguments                                                | 12.11.0 | source | %Change |
|------------|----------------------------------------------------------|---------|--------|---------|
| download   | --size 10240 --parallel 64                               | 1,974   | 1,881  | -4.71%  |
| download   | --size 10485760 --parallel 32                            | 147.6   | 141.7  | -4.00%  |
| download   | --size 1073741824 --parallel 1 --warmup 60 --duration 60 | 0.1398  | 0.1461 | 4.51%   |
| download   | --size 1073741824 --parallel 8 --warmup 60 --duration 60 | 1.069   | 1.092  | 2.15%   |


## Validation Build
https://dev.azure.com/azure-sdk/internal/_build/results?buildId=1743984&view=results